Move 'parseError' to request lib
