+++ server/sonar-web/src/main/js/apps/code/utils.ts
@@ -28,7 +28,6 @@ import {
 } from './bucket';
 import { Breadcrumb, Component } from './types';
 import { getChildren, getComponent, getBreadcrumbs } from '../../api/components';
-import { translate } from '../../helpers/l10n';
 
 const METRICS = [
   'ncloc',
@@ -235,18 +234,4 @@ export function loadMoreChildren(
       storeChildrenBreadcrumbs(componentKey, r.components);
       return r;
     });
-}
-
-/** Parse response of failed request */
-export function parseError(error: { response: Response }): Promise<string> {
-  const DEFAULT_MESSAGE = translate('default_error_message');
-
-  try {
-    return error.response
-      .json()
-      .then(r => r.errors.map((error: any) => error.msg).join('. '))
-      .catch(() => DEFAULT_MESSAGE);
-  } catch (ex) {
-    return Promise.resolve(DEFAULT_MESSAGE);
-  }
 }
