+++ server/sonar-web/src/main/js/apps/quality-gates/views/gate-conditions-delete-view.js
@@ -21,7 +21,7 @@ import ModalForm from '../../../components/common/modal-form';
 import Template from '../templates/quality-gates-condition-delete.hbs';
 import { deleteCondition } from '../../../api/quality-gates';
 import { getLocalizedMetricName } from '../../../helpers/l10n';
-import { parseError } from '../../code/utils';
+import { parseError } from '../../../helpers/request';
 
 export default ModalForm.extend({
   template: Template,
