+++ server/sonar-web/src/main/js/apps/settings/components/EmailForm.js
@@ -21,7 +21,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { translate, translateWithParameters } from '../../../helpers/l10n';
 import { sendTestEmail } from '../../../api/settings';
-import { parseError } from '../../code/utils';
+import { parseError } from '../../../helpers/request';
 import { getCurrentUser } from '../../../store/rootReducer';
 
 class EmailForm extends React.PureComponent {
